requirements.txt
discord.py[voice]
libnacl
asyncio~=3.4.3
aiohttp~=3.6.2
pillow~=7.2.0
mysql-connector-python~=8.0.18
celery[redis]~=4.4.7
PyNaCl
youtube-dl
apscheduler~=3.6.0
flask~=1.1.2
flask-restplus
Werkzeug==0.16.1
Wavelink~=0.9.6
fuzzywuzzy~=0.18.0
sqlalchemy~=1.3.19
discord-pretty-help